Protect contents from cross-origin attacks
Using Spectre, other websites can extract the contents of certain
responses even if the CORS policy doesn't allow it.

Use Cross-Origin Resource Policy (CORP) to globally opt into
Cross-Origin Read Blocking (CORB), which blocks the contents from being
delivered to other origins at all. By default, CORB only protects HTML
and JSON, under the expectation that other content types are gated by
credentials or CSRF tokens, but that's not the case here (and I can't
edit the JS to make it pass tokens).

An alternative could have been to check the Origin header, but it looks
like it's not even clear if it's supposed to be sent with cross-origin
GET fetches.

While at it, also throw in the new fancy Cross Origin Opener Policy
(COOP) to isolate the document from other origins that might open it in
a new window. I think Cross Origin Embedder Policy (COEP) is not
relevant to us because it's embedder (attacker) side.

This really should be easier than this, but also CPUs should work.